requirements.txt
flask
python-dotenv
psycopg2-binary
Flask-SQLAlchemy
flask-marshmallow
marshmallow-sqlalchemy
